requirements.txt
openai
fastapi
uvicorn
pydantic
python-dotenv
bs4